ScriptPage: Replaced pykdeuic4 generated code with cleaned up original ui file. Merged with UI_ScriptPage class and moved everything to scriptpage.py
