+++ lib/autokey/qtui/scriptpage.py
@@ -1,61 +1,158 @@
-#!/usr/bin/env python
-# coding=UTF-8
-#
-# Generated by pykdeuic4 from scriptpage.ui on Sun Mar  4 11:39:40 2012
-#
-# WARNING! All changes to this file will be lost.
-from PyKDE4 import kdecore
-from PyKDE4 import kdeui
-from PyQt4 import QtCore, QtGui
+# Copyright (C) 2011 Chris Dekter
+# Copyright (C) 2018 Thomas Hess <thomas.hess@udo.edu>
 
-try:
-    _fromUtf8 = QtCore.QString.fromUtf8
-except AttributeError:
-    _fromUtf8 = lambda s: s
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
 
-class Ui_ScriptPage(object):
-    def setupUi(self, ScriptPage):
-        ScriptPage.setObjectName(_fromUtf8("ScriptPage"))
-        ScriptPage.resize(587, 581)
-        self.verticalLayout_2 = QtGui.QVBoxLayout(ScriptPage)
-        self.verticalLayout_2.setObjectName(_fromUtf8("verticalLayout_2"))
-        self.urlLabel = KUrlLabel(ScriptPage)
-        self.urlLabel.setAlignment(QtCore.Qt.AlignCenter)
-        self.urlLabel.setObjectName(_fromUtf8("urlLabel"))
-        self.verticalLayout_2.addWidget(self.urlLabel)
-        self.scriptCodeEditor = Qsci.QsciScintilla(ScriptPage)
-        self.scriptCodeEditor.setToolTip(_fromUtf8(""))
-        self.scriptCodeEditor.setWhatsThis(_fromUtf8(""))
-        self.scriptCodeEditor.setObjectName(_fromUtf8("scriptCodeEditor"))
-        self.verticalLayout_2.addWidget(self.scriptCodeEditor)
-        self.settingsGroupbox = QtGui.QGroupBox(ScriptPage)
-        self.settingsGroupbox.setObjectName(_fromUtf8("settingsGroupbox"))
-        self.verticalLayout = QtGui.QVBoxLayout(self.settingsGroupbox)
-        self.verticalLayout.setObjectName(_fromUtf8("verticalLayout"))
-        self.promptCheckbox = QtGui.QCheckBox(self.settingsGroupbox)
-        self.promptCheckbox.setObjectName(_fromUtf8("promptCheckbox"))
-        self.verticalLayout.addWidget(self.promptCheckbox)
-        self.showInTrayCheckbox = QtGui.QCheckBox(self.settingsGroupbox)
-        self.showInTrayCheckbox.setObjectName(_fromUtf8("showInTrayCheckbox"))
-        self.verticalLayout.addWidget(self.showInTrayCheckbox)
-        self.kseparator = KSeparator(self.settingsGroupbox)
-        self.kseparator.setObjectName(_fromUtf8("kseparator"))
-        self.verticalLayout.addWidget(self.kseparator)
-        self.settingsWidget = SettingsWidget(self.settingsGroupbox)
-        self.settingsWidget.setObjectName(_fromUtf8("settingsWidget"))
-        self.verticalLayout.addWidget(self.settingsWidget)
-        self.verticalLayout_2.addWidget(self.settingsGroupbox)
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
 
-        self.retranslateUi(ScriptPage)
-        QtCore.QMetaObject.connectSlotsByName(ScriptPage)
+# You should have received a copy of the GNU General Public License
+# along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-    def retranslateUi(self, ScriptPage):
-        ScriptPage.setWindowTitle(kdecore.i18n(_fromUtf8("Form")))
-        self.urlLabel.setTipText(kdecore.i18n(_fromUtf8("Open the script in the default text editor")))
-        self.settingsGroupbox.setTitle(kdecore.i18n(_fromUtf8("Script Settings")))
-        self.promptCheckbox.setText(kdecore.i18n(_fromUtf8("Always prompt before running this script")))
-        self.showInTrayCheckbox.setText(kdecore.i18n(_fromUtf8("Show in notification icon menu")))
+import os.path
+import subprocess
 
 from PyQt4 import Qsci
-from PyKDE4.kdeui import KUrlLabel, KSeparator
-from .settingswidget import SettingsWidget
+from PyQt4.QtGui import QMessageBox
+
+from .common import inherits_from_ui_file_with_name, set_url_label, EMPTY_FIELD_REGEX, monospace_font
+
+API_FILE = os.path.join(os.path.dirname(os.path.abspath(__file__)), "data/api.txt")
+
+PROBLEM_MSG_PRIMARY = "Some problems were found"
+PROBLEM_MSG_SECONDARY = "{}\n\nYour changes have not been saved."
+
+
+class ScriptPage(*inherits_from_ui_file_with_name("scriptpage")):
+
+    def __init__(self):
+        super(ScriptPage, self).__init__()
+        self.setupUi(self)
+
+        self.scriptCodeEditor.setUtf8(1)
+
+        lex = Qsci.QsciLexerPython(self)
+        api = Qsci.QsciAPIs(lex)
+        api.load(API_FILE)
+        api.prepare()
+        self.current_script = None
+        self.scriptCodeEditor.setLexer(lex)
+
+        self.scriptCodeEditor.setBraceMatching(Qsci.QsciScintilla.SloppyBraceMatch)
+        self.scriptCodeEditor.setAutoIndent(True)
+        self.scriptCodeEditor.setBackspaceUnindents(True)
+        self.scriptCodeEditor.setIndentationWidth(4)
+        self.scriptCodeEditor.setIndentationGuides(True)
+        self.scriptCodeEditor.setIndentationsUseTabs(False)
+        self.scriptCodeEditor.setAutoCompletionThreshold(3)
+        self.scriptCodeEditor.setAutoCompletionSource(Qsci.QsciScintilla.AcsAll)
+        self.scriptCodeEditor.setCallTipsStyle(Qsci.QsciScintilla.CallTipsNoContext)
+        lex.setFont(monospace_font())
+
+    def load(self, script):
+        self.current_script = script
+        self.scriptCodeEditor.clear()
+        self.scriptCodeEditor.append(script.code)
+        self.showInTrayCheckbox.setChecked(script.showInTrayMenu)
+        self.promptCheckbox.setChecked(script.prompt)
+        self.settingsWidget.load(script)
+        self.topLevelWidget().set_undo_available(False)
+        self.topLevelWidget().set_redo_available(False)
+
+        if self.is_new_item():
+            self.urlLabel.setEnabled(False)
+            self.urlLabel.setText("(Unsaved)")  # TODO: i18n
+        else:
+            set_url_label(self.urlLabel, self.current_script.path)
+
+    def save(self):
+        self.settingsWidget.save()
+        self.current_script.code = str(self.scriptCodeEditor.text())
+        self.current_script.showInTrayMenu = self.showInTrayCheckbox.isChecked()
+
+        self.current_script.persist()
+        set_url_label(self.urlLabel, self.current_script.path)
+        return False
+
+    def set_item_title(self, title):
+        self.current_script.description = title
+
+    def rebuild_item_path(self):
+        self.current_script.rebuild_path()
+
+    def is_new_item(self):
+        return self.current_script.path is None
+
+    def reset(self):
+        self.load(self.current_script)
+        self.topLevelWidget().set_undo_available(False)
+        self.topLevelWidget().set_redo_available(False)
+
+    def set_dirty(self):
+        self.topLevelWidget().set_dirty()
+
+    def start_record(self):
+        self.scriptCodeEditor.append("\n")
+
+    def start_key_sequence(self):
+        self.scriptCodeEditor.append("keyboard.send_keys(\"")
+
+    def end_key_sequence(self):
+        self.scriptCodeEditor.append("\")\n")
+
+    def append_key(self, key):
+        self.scriptCodeEditor.append(key)
+
+    def append_hotkey(self, key, modifiers):
+        keyString = self.current_script.get_hotkey_string(key, modifiers)
+        self.scriptCodeEditor.append(keyString)
+
+    def append_mouseclick(self, xCoord, yCoord, button, windowTitle):
+        self.scriptCodeEditor.append("mouse.click_relative(%d, %d, %d) # %s\n" % (xCoord, yCoord, int(button), windowTitle))
+
+    def undo(self):
+        self.scriptCodeEditor.undo()
+        self.topLevelWidget().set_undo_available(self.scriptCodeEditor.isUndoAvailable())
+
+    def redo(self):
+        self.scriptCodeEditor.redo()
+        self.topLevelWidget().set_redo_available(self.scriptCodeEditor.isRedoAvailable())
+
+    def validate(self):
+        errors = []
+
+        # Check script code
+        code = str(self.scriptCodeEditor.text())
+        if EMPTY_FIELD_REGEX.match(code):
+            errors.append("The script code can't be empty")  # TODO: i18n
+
+        # Check settings
+        errors += self.settingsWidget.validate()
+
+        if errors:
+            msg = PROBLEM_MSG_SECONDARY.format('\n'.join([str(e) for e in errors]))
+            header = PROBLEM_MSG_PRIMARY
+            QMessageBox.critical(self.topLevelWidget(), header, msg)
+
+        return not bool(errors)
+
+    # --- Signal handlers
+
+    def on_scriptCodeEditor_textChanged(self):
+        self.set_dirty()
+        self.topLevelWidget().set_undo_available(self.scriptCodeEditor.isUndoAvailable())
+        self.topLevelWidget().set_redo_available(self.scriptCodeEditor.isRedoAvailable())
+
+    def on_promptCheckbox_stateChanged(self, state):
+        self.set_dirty()
+
+    def on_showInTrayCheckbox_stateChanged(self, state):
+        self.set_dirty()
+
+    def on_urlLabel_leftClickedUrl(self, url=None):
+        if url: subprocess.Popen(["/usr/bin/xdg-open", url])
